split python downloader entrypoints
