+++ Palladium/Services/Python/palladium_ytdlp/entrypoints.py
@@ -1,1259 +1,12 @@
-import contextlib
-import io
-import importlib
-import json
-import logging
-import os
-import re
-import runpy
-import sys
-import traceback
-import urllib.parse
-
-from .args import (
-    build_preset_args,
-    has_custom_output_template,
-    parse_custom_args,
-    parse_extra_args,
-    parse_preset_args_map,
-    strip_checkbox_owned_download_args,
-)
-from .ffmpeg_bridge import (
-    SwiftFFmpegBridge,
-    YTDLPFFmpegBridgeAdapter,
-    is_cancel_requested,
-    patch_ytdlp_for_swiftffmpeg,
-)
-from .files import cleanup_temp_download_files, detect_downloaded_files, has_primary_media_file
-from .packages import (
-    build_package_install_plan,
-    build_pip_install_args,
-    check_package_updates,
-    cleanup_target_package,
-    collect_versions,
-    ensure_pip_entrypoint,
-    fetch_package_index_versions,
-    is_package_installed,
-    parse_package_source,
-)
-from .shared import TRACKED_PACKAGES, TailBuffer, Tee, open_live_log_stream
-from .webkit_jsi import ensure_safe_webkit_jsi_runtime
-
-PLAYLIST_PROGRESS_PREFIX = "[palladium][playlist-progress] "
-
-
-class GalleryDLOutputCapture(io.StringIO):
-    """String output capture compatible with gallery-dl's stream setup."""
-
-    def reconfigure(self, **_options):
-        return None
-
-
-def reset_gallery_dl_runtime():
-    """Discard gallery-dl state and log handlers from a previous in-process run."""
-    loggers = [logging.getLogger()]
-    loggers.extend(
-        logger
-        for logger in logging.Logger.manager.loggerDict.values()
-        if isinstance(logger, logging.Logger)
-    )
-    for logger in loggers:
-        for handler in list(logger.handlers):
-            if not isinstance(getattr(handler, "stream", None), (Tee, GalleryDLOutputCapture)):
-                continue
-            logger.removeHandler(handler)
-            handler.close()
-
-    for module_name in list(sys.modules):
-        if module_name == "gallery_dl" or module_name.startswith("gallery_dl."):
-            del sys.modules[module_name]
-
-
-def run_gallery_dl_module():
-    reset_gallery_dl_runtime()
-    try:
-        runpy.run_module("gallery_dl", run_name="__main__", alter_sys=True)
-    finally:
-        reset_gallery_dl_runtime()
-
-
-def invalidate_runtime_package_modules():
-    prefixes = ("yt_dlp", "yt_dlp_plugins")
-    webkit_jsi_loaded = any(
-        name == "yt_dlp_plugins.webkit_jsi" or name.startswith("yt_dlp_plugins.webkit_jsi.")
-        for name in sys.modules
-    )
-    if webkit_jsi_loaded:
-        print("[palladium] restart required to refresh loaded webkit jsi runtime")
-        return True
-
-    stale_modules = [
-        name
-        for name in sys.modules
-        if any(name == prefix or name.startswith(f"{prefix}.") for prefix in prefixes)
-    ]
-
-    for name in stale_modules:
-        sys.modules.pop(name, None)
-
-    importlib.invalidate_caches()
-    print(f"[palladium] invalidated {len(stale_modules)} cached yt-dlp runtime module(s)")
-    return False
-
-
-class PlaylistProgressCollector:
-    playlist_title_pattern = re.compile(r"^\[download\] Downloading playlist: (?P<title>.+)$")
-    playlist_count_pattern = re.compile(
-        r"^\[(?:youtube:tab|download)\].*Downloading (?P<count>\d+) items? of (?P<total>\d+)$"
-    )
-    playlist_item_pattern = re.compile(r"^\[download\] Downloading item (?P<index>\d+) of (?P<total>\d+)$")
-    item_url_pattern = re.compile(r"^\[[^\]]+\] Extracting URL: (?P<url>.+)$")
-    destination_pattern = re.compile(r"^\[(?:download|ExtractAudio)\] Destination: (?P<path>.+)$")
-
-    def __init__(self, live_log_stream=None):
-        self.live_log_stream = live_log_stream
-        self.pending_line = ""
-        self.playlist_title = None
-        self.playlist_expected_count = None
-        self.playlist_completed_count = 0
-        self.playlist_failed_count = 0
-        self.playlist_failed_items = []
-        self.failed_item_records = []
-        self.current_item_index = None
-        self.current_item_title = None
-        self.current_item_url = None
-        self.current_item_paths = set()
-        self.current_item_had_error = False
-        self.current_item_error_line = None
-        self.is_playlist_run = False
-        self.last_emitted_payload = None
-        self.suspend_tracking = False
-
-    def write(self, data):
-        text = str(data)
-        if not text:
-            return 0
-
-        if self.suspend_tracking:
-            return len(text)
-
-        normalized = text.replace("\r\n", "\n").replace("\r", "\n")
-        combined = self.pending_line + normalized
-        lines = combined.split("\n")
-        if normalized.endswith("\n"):
-            self.pending_line = ""
-        else:
-            self.pending_line = lines.pop() if lines else combined
-
-        for line in lines:
-            self._handle_line(line.strip())
-
-        return len(text)
-
-    def flush(self):
-        return None
-
-    def finalize(self, cancelled=False):
-        if self.pending_line:
-            self._handle_line(self.pending_line.strip())
-            self.pending_line = ""
-        self._finalize_current_item(aborted=cancelled)
-
-    def result_kind(self, success, cancelled):
-        if cancelled:
-            return "cancelled"
-
-        if not self.is_playlist_run:
-            return "success" if success else "error"
-
-        completed = self.playlist_completed_count
-        failed = self.playlist_failed_count
-
-        if completed > 0 and failed > 0:
-            return "partial"
-        if completed > 0 and failed == 0:
-            return "success"
-        return "error"
-
-    def snapshot(self, result_kind=None):
-        return {
-            "playlist_title": self.playlist_title,
-            "playlist_expected_count": self.playlist_expected_count,
-            "playlist_completed_count": self.playlist_completed_count,
-            "playlist_failed_count": self.playlist_failed_count,
-            "playlist_failed_items": list(self.playlist_failed_items),
-            "current_item_index": self.current_item_index,
-            "current_item_title": self.current_item_title,
-            "result_kind": result_kind,
-        }
-
-    def _handle_line(self, line):
-        if not line:
-            return
-
-        playlist_title_match = self.playlist_title_pattern.match(line)
-        if playlist_title_match:
-            self.is_playlist_run = True
-            self.playlist_title = playlist_title_match.group("title").strip() or None
-            self._emit()
-            return
-
-        playlist_count_match = self.playlist_count_pattern.match(line)
-        if playlist_count_match:
-            self.is_playlist_run = True
-            self.playlist_expected_count = int(playlist_count_match.group("total"))
-            self._emit()
-            return
-
-        playlist_item_match = self.playlist_item_pattern.match(line)
-        if playlist_item_match:
-            self._finalize_current_item()
-            self.is_playlist_run = True
-            self.current_item_index = int(playlist_item_match.group("index"))
-            self.playlist_expected_count = int(playlist_item_match.group("total"))
-            self.current_item_title = None
-            self.current_item_url = None
-            self.current_item_paths = set()
-            self.current_item_had_error = False
-            self.current_item_error_line = None
-            self._emit()
-            return
-
-        item_url_match = self.item_url_pattern.match(line)
-        if item_url_match and self.current_item_index is not None:
-            self.current_item_url = item_url_match.group("url").strip() or None
-            return
-
-        destination_match = self.destination_pattern.match(line)
-        if destination_match and self.current_item_index is not None:
-            path = destination_match.group("path").strip()
-            if path:
-                self.current_item_paths.add(path)
-                guessed_title = self._title_from_path(path)
-                if guessed_title:
-                    self.current_item_title = guessed_title
-                    self._emit()
-            return
-
-        if line.startswith("ERROR:") and self.current_item_index is not None:
-            self.current_item_had_error = True
-            self.current_item_error_line = line
-            self._emit()
-
-    def _finalize_current_item(self, aborted=False):
-        if self.current_item_index is None:
-            return
-
-        if not aborted:
-            has_output = any(self._path_exists(path) for path in self.current_item_paths)
-            if has_output:
-                self.playlist_completed_count += 1
-            else:
-                self._record_failed_item()
-
-        self.current_item_index = None
-        self.current_item_title = None
-        self.current_item_url = None
-        self.current_item_paths = set()
-        self.current_item_had_error = False
-        self.current_item_error_line = None
-        self._emit()
-
-    def retry_candidates(self):
-        return [dict(record) for record in self.failed_item_records if self._is_thumbnail_failure(record.get("error_line"))]
-
-    def mark_retry_success(self, item_index):
-        updated_records = []
-        removed_count = 0
-        for record in self.failed_item_records:
-            if record.get("index") == item_index:
-                removed_count += 1
-                continue
-            updated_records.append(record)
-
-        if removed_count == 0:
-            return
-
-        self.failed_item_records = updated_records
-        self.playlist_failed_count = max(0, self.playlist_failed_count - removed_count)
-        self.playlist_completed_count += removed_count
-        self._rebuild_failed_items()
-        self._emit()
-
-    def mark_retry_failed(self, item_index, error_line):
-        for record in self.failed_item_records:
-            if record.get("index") != item_index:
-                continue
-            if error_line:
-                record["error_line"] = error_line
-            break
-        self._rebuild_failed_items()
-        self._emit()
-
-    def _record_failed_item(self):
-        self.playlist_failed_count += 1
-        failed_label = self.current_item_title or f"Item {self.current_item_index}"
-        if self.current_item_error_line:
-            failed_label = f"{failed_label}: {self.current_item_error_line}"
-        self.playlist_failed_items.append(failed_label)
-        self.failed_item_records.append({
-            "index": self.current_item_index,
-            "title": self.current_item_title,
-            "url": self.current_item_url,
-            "error_line": self.current_item_error_line,
-        })
-
-    def _rebuild_failed_items(self):
-        labels = []
-        for record in self.failed_item_records:
-            failed_label = record.get("title") or f"Item {record.get('index')}"
-            if record.get("error_line"):
-                failed_label = f"{failed_label}: {record['error_line']}"
-            labels.append(failed_label)
-        self.playlist_failed_items = labels
-
-    def _is_thumbnail_failure(self, error_line):
-        if not error_line:
-            return False
-        lower = error_line.lower()
-        return (
-            "thumbnail" in lower
-            or "thumbnails" in lower
-            or "embed-thumbnail" in lower
-            or "convert-thumbnails" in lower
-            or ".webp" in lower
-            or ".png" in lower
-        )
-
-    def set_tracking_suspended(self, suspended):
-        self.suspend_tracking = bool(suspended)
-
-    def _emit(self):
-        if self.live_log_stream is None:
-            return
-
-        payload = self.snapshot()
-        encoded_payload = json.dumps(payload, ensure_ascii=False, sort_keys=True)
-        if encoded_payload == self.last_emitted_payload:
-            return
-
-        self.last_emitted_payload = encoded_payload
-        try:
-            self.live_log_stream.write(f"{PLAYLIST_PROGRESS_PREFIX}{encoded_payload}\n")
-            self.live_log_stream.flush()
-        except Exception:
-            pass
-
-    def _path_exists(self, path):
-        try:
-            return os.path.isfile(path) and os.path.getsize(path) > 0
-        except Exception:
-            return False
-
-    def _title_from_path(self, path):
-        file_name = os.path.splitext(os.path.basename(path))[0].strip()
-        if not file_name:
-            return None
-
-        cleaned = re.sub(r"^\d+\s+-\s+", "", file_name).strip()
-        return cleaned or file_name
-
-
-def raise_if_cancel_requested(cancel_file_path, message):
-    if is_cancel_requested(cancel_file_path):
-        print(message)
-        raise KeyboardInterrupt("cancel requested")
-
-
-def without_thumbnail_download_args(download_behavior_args):
-    result = []
-    skip_next = False
-    for arg in download_behavior_args:
-        if skip_next:
-            skip_next = False
-            continue
-        if arg == "--convert-thumbnails":
-            skip_next = True
-            continue
-        if arg == "--embed-thumbnail":
-            continue
-        result.append(arg)
-    return result
-
-
-def extract_last_error_line(output):
-    lines = [line.strip() for line in str(output).splitlines() if line.strip().startswith("ERROR:")]
-    return lines[-1] if lines else None
-
-
-def ensure_gallery_dl_installed(install_target, package_source):
-    installed, version, source = is_package_installed("gallery-dl", install_target=install_target)
-    if installed:
-        print(f"[palladium] gallery-dl already installed ({version} via {source})")
-        return False, 0
-
-    print("[palladium] gallery-dl package missing")
-    pip_main = ensure_pip_entrypoint(install_target)
-    if pip_main is None:
-        return True, 1
-
-    packages = ["gallery-dl"]
-    pip_args = build_pip_install_args(
-        packages,
-        install_target=install_target,
-        allow_prereleases=bool(package_source.get("allow_prereleases")),
-    )
-    result = pip_main(pip_args)
-    exit_code = 0 if result is None else int(result)
-    print(f"[palladium] gallery-dl pip exit code: {exit_code}")
-    return True, exit_code
-
-
-def gallery_dl_args(url, cookie_file_path=None, destination=None, selection_range=None, resolve=False):
-    args = ["gallery-dl", "--config-ignore", "--no-colors", "--verbose"]
-    if cookie_file_path:
-        if os.path.isfile(cookie_file_path):
-            args.extend(["--cookies", cookie_file_path])
-        else:
-            print(f"[palladium] cookie file missing, ignoring: {cookie_file_path}")
-    if resolve:
-        args.append("--resolve-urls")
-    if destination:
-        args.extend(["--destination", destination])
-    if selection_range:
-        args.extend(["--range", selection_range])
-    args.append(url)
-    return args
-
-
-GALLERY_AUDIO_EXTENSIONS = {
-    ".aac",
-    ".aif",
-    ".aiff",
-    ".alac",
-    ".flac",
-    ".m4a",
-    ".mp3",
-    ".oga",
-    ".ogg",
-    ".opus",
-    ".wav",
-    ".weba",
-    ".wma",
-}
-
-GALLERY_IMAGE_EXTENSIONS = {
-    ".avif",
-    ".bmp",
-    ".gif",
-    ".heic",
-    ".heif",
-    ".jpeg",
-    ".jpg",
-    ".png",
-    ".tif",
-    ".tiff",
-    ".webp",
-}
-
-
-def gallery_item_media_type(url):
-    try:
-        path = urllib.parse.unquote(urllib.parse.urlparse(url).path)
-        extension = os.path.splitext(path)[1].lower()
-    except Exception:
-        extension = ""
-    if extension in GALLERY_AUDIO_EXTENSIONS:
-        return "audio"
-    if extension in GALLERY_IMAGE_EXTENSIONS:
-        return "image"
-    return "file"
-
-
-def gallery_item_title(url, index, media_type="image"):
-    try:
-        path = urllib.parse.unquote(urllib.parse.urlparse(url).path)
-        name = os.path.basename(path)
-        if name:
-            return name
-    except Exception:
-        pass
-    if media_type == "audio":
-        return f"Audio {index}"
-    if media_type == "file":
-        return f"File {index}"
-    return f"Image {index}"
-
-
-def gallery_resolution_error_message(output):
-    """Return gallery-dl's last user-facing error line without diagnostic noise."""
-    for line in reversed(output.splitlines()):
-        match = re.match(r"^\[[^\]]+\]\[error\]\s+(.+)$", line.strip(), re.IGNORECASE)
-        if match:
-            return match.group(1)
-    return None
-
-
-def run_gallery_dl_resolver(download_url_override=None, cookie_file_path_override=None, live_log_fd_override=None, package_source_json_override=None):
-    output = TailBuffer()
-    console_stdout = sys.__stdout__ if sys.__stdout__ is not None else None
-    console_stderr = sys.__stderr__ if sys.__stderr__ is not None else None
-    live_log_stream = open_live_log_stream(live_log_fd_override)
-    install_target = os.environ.get("PALLADIUM_PYTHON_PACKAGES")
-    package_source = parse_package_source(package_source_json_override)
-    url = str(download_url_override or "").strip()
-    cookie_file_path = str(cookie_file_path_override or "").strip()
-    argv_backup = sys.argv[:]
-    items = []
-    pip_attempted = False
-    pip_exit_code = None
-    success = False
-    captured_output = ""
-    gallery_dl_exit_code = None
-
-    with contextlib.redirect_stdout(Tee(output, console_stdout, live_log_stream)), contextlib.redirect_stderr(Tee(output, console_stderr, live_log_stream)):
-        try:
-            if install_target:
-                os.makedirs(install_target, exist_ok=True)
-                if install_target not in sys.path:
-                    sys.path.insert(0, install_target)
-            if not url:
-                print("[palladium] gallery-dl resolver received no URL")
-            else:
-                pip_attempted, pip_exit_code = ensure_gallery_dl_installed(install_target, package_source)
-                if pip_exit_code == 0:
-                    captured = GalleryDLOutputCapture()
-                    sys.argv = gallery_dl_args(url, cookie_file_path=cookie_file_path, resolve=True)
-                    try:
-                        with contextlib.redirect_stdout(captured):
-                            run_gallery_dl_module()
-                    except SystemExit as exc:
-                        if exc.code not in (None, 0):
-                            gallery_dl_exit_code = exc.code
-                    finally:
-                        captured_output = captured.getvalue()
-
-                    if gallery_dl_exit_code is None:
-                        seen = set()
-                        for line in captured_output.splitlines():
-                            candidate = line.strip()
-                            if not candidate.startswith(("http://", "https://")) or candidate in seen:
-                                continue
-                            seen.add(candidate)
-                            media_type = gallery_item_media_type(candidate)
-                            items.append({
-                                "index": len(items) + 1,
-                                "url": candidate,
-                                "title": gallery_item_title(candidate, len(items) + 1, media_type),
-                                "media_type": media_type,
-                            })
-                        success = bool(items)
-                        print(f"[palladium] gallery-dl resolved {len(items)} item(s)")
-        except Exception:
-            print("[palladium] gallery-dl resolution failed")
-            traceback.print_exc()
-        finally:
-            sys.argv = argv_backup
-            if live_log_stream is not None:
-                live_log_stream.flush()
-
-    error_message = None
-    if not success:
-        error_message = gallery_resolution_error_message("\n".join((captured_output, output.getvalue())))
-
-    return json.dumps({
-        "success": success,
-        "pip_attempted": pip_attempted,
-        "pip_exit_code": pip_exit_code,
-        "items": items,
-        "output": output.getvalue(),
-        "error_message": error_message,
-    })
-
-
-def run_gallery_dl_flow(download_url_override=None, selection_range_override=None, cookie_file_path_override=None, run_output_dir_override=None, live_log_fd_override=None, package_source_json_override=None):
-    output = TailBuffer()
-    console_stdout = sys.__stdout__ if sys.__stdout__ is not None else None
-    console_stderr = sys.__stderr__ if sys.__stderr__ is not None else None
-    live_log_stream = open_live_log_stream(live_log_fd_override)
-    install_target = os.environ.get("PALLADIUM_PYTHON_PACKAGES")
-    url = str(download_url_override or "").strip()
-    selection_range = str(selection_range_override or "").strip()
-    cookie_file_path = str(cookie_file_path_override or "").strip()
-    run_output_dir = str(run_output_dir_override or "").strip()
-    package_source = parse_package_source(package_source_json_override)
-    argv_backup = sys.argv[:]
-    pip_attempted = False
-    pip_exit_code = None
-    exit_code = 1
-    cancelled = False
-    downloaded_paths = []
-    primary_downloaded_path = None
-
-    with contextlib.redirect_stdout(Tee(output, console_stdout, live_log_stream)), contextlib.redirect_stderr(Tee(output, console_stderr, live_log_stream)):
-        try:
-            if install_target:
-                os.makedirs(install_target, exist_ok=True)
-                if install_target not in sys.path:
-                    sys.path.insert(0, install_target)
-            if run_output_dir:
-                os.makedirs(run_output_dir, exist_ok=True)
-                cleanup_temp_download_files(run_output_dir)
-            pip_attempted, pip_exit_code = ensure_gallery_dl_installed(install_target, package_source)
-            if pip_exit_code == 0 and url and selection_range:
-                sys.argv = gallery_dl_args(
-                    url,
-                    cookie_file_path,
-                    run_output_dir,
-                    selection_range,
-                )
-                print(f"[palladium] running gallery-dl for selected range: {selection_range}")
-                try:
-                    run_gallery_dl_module()
-                    exit_code = 0
-                except KeyboardInterrupt:
-                    cancelled = True
-                    exit_code = 130
-                except SystemExit as exc:
-                    exit_code = 0 if exc.code in (None, 0) else int(exc.code)
-            else:
-                print("[palladium] gallery-dl download missing URL or selected images")
-
-            if not cancelled and run_output_dir:
-                downloaded_paths, primary_downloaded_path = detect_downloaded_files(run_output_dir)
-                if downloaded_paths and exit_code != 0:
-                    exit_code = 0
-        except KeyboardInterrupt:
-            cancelled = True
-            exit_code = 130
-        except Exception:
-            print("[palladium] gallery-dl download failed")
-            traceback.print_exc()
-        finally:
-            sys.argv = argv_backup
-            if live_log_stream is not None:
-                live_log_stream.flush()
-
-    success = pip_exit_code in (None, 0) and exit_code == 0 and bool(downloaded_paths) and not cancelled
-    return json.dumps({
-        "pip_attempted": pip_attempted,
-        "pip_exit_code": pip_exit_code,
-        "yt_exit_code": exit_code,
-        "cancelled": cancelled,
-        "success": success,
-        "downloaded_paths": downloaded_paths,
-        "primary_downloaded_path": primary_downloaded_path,
-        "downloaded_path": primary_downloaded_path,
-        "output": output.getvalue(),
-        "result_kind": "cancelled" if cancelled else ("success" if success else "error"),
-    })
-
-
-def run_retry_without_thumbnails(
-    retry_candidate,
-    download_url,
-    run_output_dir,
-    cache_dir,
-    download_playlist,
-    output_args,
-    download_behavior_args,
-    preset_args,
-    extra_args,
-    bridge_adapter,
-):
-    item_index = retry_candidate.get("index")
-    if item_index is None:
-        return False, "ERROR: retry skipped because playlist item index was missing"
-
-    retry_behavior_args = without_thumbnail_download_args(download_behavior_args)
-    retry_url = download_url
-    retry_args = [
-        "yt-dlp",
-        "-v",
-        "--no-check-certificate",
-        "--remote-components",
-        "ejs:github",
-        "--cache-dir",
-        cache_dir if cache_dir else os.path.join(".", ".cache"),
-        "--force-overwrites",
-        "--no-continue",
-        "-P",
-        run_output_dir if run_output_dir else ".",
-        *output_args,
-        *retry_behavior_args,
-        *preset_args,
-        *extra_args,
-    ]
-
-    if download_playlist:
-        retry_args.extend(["--playlist-items", str(item_index)])
-    retry_args.append(retry_url)
-
-    sys.argv = retry_args
-    try:
-        with patch_ytdlp_for_swiftffmpeg(bridge_adapter):
-            runpy.run_module("yt_dlp", run_name="__main__", alter_sys=True)
-        return True, None
-    except SystemExit as exc:
-        if exc.code in (None, 0):
-            return True, None
-        return False, f"ERROR: retry without thumbnails failed with exit code {exc.code}"
-    except Exception:
-        traceback.print_exc()
-        return False, extract_last_error_line(traceback.format_exc()) or "ERROR: retry without thumbnails failed"
-
-
-def run_yt_dlp_flow(
-    download_url_override=None,
-    download_preset_override=None,
-    preset_args_json_override=None,
-    extra_args_override=None,
-    download_playlist_override=None,
-    download_subtitles_override=None,
-    embed_thumbnail_override=None,
-    auto_retry_failed_downloads_override=None,
-    subtitle_language_pattern_override=None,
-    cookie_file_path_override=None,
-    run_output_dir_override=None,
-    live_log_fd_override=None,
-    package_source_json_override=None,
-):
-    output = TailBuffer()
-    console_stdout = sys.__stdout__ if sys.__stdout__ is not None else None
-    console_stderr = sys.__stderr__ if sys.__stderr__ is not None else None
-    pip_attempted = False
-    pip_exit_code = None
-    yt_exit_code = None
-    downloaded_paths = []
-    primary_downloaded_path = None
-    cancelled = False
-    success = False
-    if download_url_override is None:
-        download_url = os.environ.get("PALLADIUM_DOWNLOAD_URL", "").strip()
-    else:
-        download_url = str(download_url_override).strip()
-
-    if download_preset_override is None:
-        download_preset = os.environ.get("PALLADIUM_DOWNLOAD_PRESET", "auto_video").strip()
-    else:
-        download_preset = str(download_preset_override).strip()
-    if preset_args_json_override is None:
-        preset_args_json = os.environ.get("PALLADIUM_PRESET_ARGS_JSON", "").strip()
-    else:
-        preset_args_json = str(preset_args_json_override).strip()
-    if extra_args_override is None:
-        extra_args_text = os.environ.get("PALLADIUM_EXTRA_ARGS", "").strip()
-    else:
-        extra_args_text = str(extra_args_override).strip()
-    if download_playlist_override is None:
-        download_playlist = os.environ.get("PALLADIUM_DOWNLOAD_PLAYLIST", "").strip().lower() in ("1", "true", "yes", "on")
-    else:
-        download_playlist = bool(download_playlist_override)
-    if download_subtitles_override is None:
-        download_subtitles = os.environ.get("PALLADIUM_DOWNLOAD_SUBTITLES", "").strip().lower() in ("1", "true", "yes", "on")
-    else:
-        download_subtitles = bool(download_subtitles_override)
-    if embed_thumbnail_override is None:
-        embed_thumbnail = os.environ.get("PALLADIUM_EMBED_THUMBNAIL", "").strip().lower() in ("1", "true", "yes", "on")
-    else:
-        embed_thumbnail = bool(embed_thumbnail_override)
-    if auto_retry_failed_downloads_override is None:
-        auto_retry_failed_downloads = os.environ.get("PALLADIUM_AUTO_RETRY_FAILED_DOWNLOADS", "").strip().lower() in ("1", "true", "yes", "on")
-    else:
-        auto_retry_failed_downloads = bool(auto_retry_failed_downloads_override)
-    if subtitle_language_pattern_override is None:
-        subtitle_language_pattern = os.environ.get("PALLADIUM_SUBTITLE_LANGUAGE_PATTERN", "en").strip() or "en"
-    else:
-        subtitle_language_pattern = str(subtitle_language_pattern_override).strip() or "en"
-    if subtitle_language_pattern == "en.*":
-        subtitle_language_pattern = "en"
-    if cookie_file_path_override is None:
-        cookie_file_path = os.environ.get("PALLADIUM_COOKIE_FILE_PATH", "").strip()
-    else:
-        cookie_file_path = str(cookie_file_path_override).strip()
-    downloads_dir = os.environ.get("PALLADIUM_DOWNLOADS", "").strip()
-    if run_output_dir_override is None:
-        run_output_dir = os.environ.get("PALLADIUM_RUN_OUTPUT_DIR", "").strip() or downloads_dir
-    else:
-        run_output_dir = str(run_output_dir_override).strip() or downloads_dir
-    install_target = os.environ.get("PALLADIUM_PYTHON_PACKAGES")
-    cache_dir = os.environ.get("PALLADIUM_CACHE_DIR", "").strip()
-    cancel_file_path = os.environ.get("PALLADIUM_CANCEL_FILE", "").strip()
-    live_log_stream = open_live_log_stream(live_log_fd_override)
-    package_source = parse_package_source(package_source_json_override)
-    playlist_progress = PlaylistProgressCollector(live_log_stream=live_log_stream)
-
-    with contextlib.redirect_stdout(Tee(output, console_stdout, live_log_stream, playlist_progress)), contextlib.redirect_stderr(Tee(output, console_stderr, live_log_stream, playlist_progress)):
-        argv_backup = sys.argv[:]
-        cwd_backup = os.getcwd()
-        try:
-            os.environ["PYTHONIOENCODING"] = "utf-8"
-            if install_target:
-                os.makedirs(install_target, exist_ok=True)
-                if install_target not in sys.path:
-                    sys.path.insert(0, install_target)
-                print(f"[palladium] package install target: {install_target}")
-            if downloads_dir:
-                os.makedirs(downloads_dir, exist_ok=True)
-                print(f"[palladium] download target: {downloads_dir}")
-            if run_output_dir:
-                os.makedirs(run_output_dir, exist_ok=True)
-                print(f"[palladium] run output target: {run_output_dir}")
-            if cache_dir:
-                os.makedirs(cache_dir, exist_ok=True)
-                print(f"[palladium] cache target: {cache_dir}")
-
-            needs_yt_dlp_install = False
-            needs_webkit_jsi_install = False
-
-            print("[palladium] checking yt-dlp package metadata")
-            yt_dlp_installed, yt_dlp_version, yt_dlp_source = is_package_installed(
-                "yt-dlp",
-                install_target=install_target,
-            )
-            if yt_dlp_installed:
-                print(f"[palladium] yt-dlp already installed ({yt_dlp_version} via {yt_dlp_source})")
-            else:
-                needs_yt_dlp_install = True
-                print("[palladium] yt-dlp package missing")
-
-            print("[palladium] checking yt-dlp-apple-webkit-jsi package metadata")
-            webkit_jsi_installed, webkit_jsi_version, webkit_jsi_source = is_package_installed(
-                "yt-dlp-apple-webkit-jsi",
-                install_target=install_target,
-            )
-            if webkit_jsi_installed:
-                print(f"[palladium] yt-dlp-apple-webkit-jsi already installed ({webkit_jsi_version} via {webkit_jsi_source})")
-            else:
-                needs_webkit_jsi_install = True
-                print("[palladium] yt-dlp-apple-webkit-jsi missing")
-
-            if needs_yt_dlp_install or needs_webkit_jsi_install:
-                raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before pip install")
-                pip_attempted = True
-                pip_main = ensure_pip_entrypoint(install_target)
-                if pip_main is not None:
-                    packages = []
-                    if package_source.get("mode") == "custom":
-                        packages = list(package_source.get("custom_specs") or [])
-                    else:
-                        if needs_yt_dlp_install:
-                            packages.append("yt-dlp")
-                        if needs_webkit_jsi_install:
-                            packages.append("yt-dlp-apple-webkit-jsi")
-
-                    try:
-                        raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before pip install")
-                        pip_args = build_pip_install_args(
-                            packages,
-                            install_target=install_target,
-                            allow_prereleases=bool(package_source.get("allow_prereleases")),
-                        )
-                        if packages:
-                            pip_result = pip_main(pip_args)
-                            pip_exit_code = 0 if pip_result is None else int(pip_result)
-                            print(f"[palladium] pip exit code: {pip_exit_code}")
-                        else:
-                            pip_exit_code = 1
-                            print("[palladium] no package specs available for install")
-                    except Exception:
-                        pip_exit_code = 1
-                        print("[palladium] pip install failed")
-                        traceback.print_exc()
-                else:
-                    pip_exit_code = 1
-
-                installed_versions = collect_versions(install_target=install_target, allow_cache_fallback=False)
-                print(f"[palladium] yt-dlp after install: {installed_versions.get('yt-dlp', 'not installed')}")
-                print(
-                    "[palladium] yt-dlp-apple-webkit-jsi after install: "
-                    f"{installed_versions.get('yt-dlp-apple-webkit-jsi', 'not installed')}"
-                )
-
-            if package_source.get("skip_webkit_patch"):
-                print("[palladium] skipping webkit patch by configuration")
-            else:
-                raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before webkit patch")
-                ensure_safe_webkit_jsi_runtime(install_target)
-
-            if not download_url:
-                print("[palladium] no URL provided")
-                yt_exit_code = 1
-            elif is_cancel_requested(cancel_file_path):
-                print("[palladium] cancellation requested before run")
-                cancelled = True
-                yt_exit_code = 130
-            else:
-                print(f"[palladium] running yt-dlp -v {download_url}")
-
-            if download_url:
-                raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before bridge startup")
-                bridge = None
-                bridge_adapter = None
-                try:
-                    bridge = SwiftFFmpegBridge()
-                    bridge_adapter = YTDLPFFmpegBridgeAdapter(bridge, cancel_file_path=cancel_file_path)
-                    print("[palladium][ffmpeg-bridge] bridge loaded")
-                    print("[palladium][ffmpeg-bridge] startup probes skipped")
-                except Exception as bridge_error:
-                    print(f"[palladium] swift ffmpeg bridge error: {bridge_error}")
-                    traceback.print_exc()
-                    yt_exit_code = 1
-
-                if yt_exit_code != 1:
-                    if run_output_dir:
-                        os.chdir(run_output_dir)
-                    raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before download cleanup")
-                    cleanup_temp_download_files(run_output_dir)
-
-                    if is_cancel_requested(cancel_file_path):
-                        print("[palladium] cancellation requested before yt-dlp start")
-                        cancelled = True
-                        yt_exit_code = 130
-
-                    if yt_exit_code is None:
-                        preset_args_map = parse_preset_args_map(preset_args_json)
-                        selected_args = preset_args_map.get(download_preset, "")
-                        if selected_args:
-                            preset_args = parse_custom_args(selected_args)
-                            print(f"[palladium] preset args override: {download_preset}")
-                        elif download_preset == "custom":
-                            preset_args = []
-                            print("[palladium] preset: custom (no args)")
-                        else:
-                            preset_args = build_preset_args(download_preset)
-                        preset_args = strip_checkbox_owned_download_args(preset_args)
-                        extra_args = strip_checkbox_owned_download_args(parse_extra_args(extra_args_text))
-                        raise_if_cancel_requested(cancel_file_path, "[palladium] cancellation requested before yt-dlp invocation")
-                        output_args = []
-                        download_behavior_args = []
-                        if has_custom_output_template(preset_args) or has_custom_output_template(extra_args):
-                            print("[palladium] custom output template detected")
-                        else:
-                            if download_playlist:
-                                output_args = ["-o", "%(playlist_index)03d - %(title)s.%(ext)s"]
-                            else:
-                                output_args = ["-o", "%(title)s.%(ext)s"]
-
-                        if not download_playlist:
-                            download_behavior_args.append("--no-playlist")
-
-                        if download_subtitles:
-                            download_behavior_args.extend(["--write-subs", "--write-auto-subs"])
-                            if subtitle_language_pattern == "all":
-                                download_behavior_args.append("--all-subs")
-                            else:
-                                download_behavior_args.extend(["--sub-langs", subtitle_language_pattern])
-
-                        if embed_thumbnail:
-                            download_behavior_args.extend(["--convert-thumbnails", "png", "--embed-thumbnail"])
-
-                        if cookie_file_path:
-                            if os.path.isfile(cookie_file_path):
-                                print(f"[palladium] using cookie file: {cookie_file_path}")
-                                download_behavior_args.extend(["--cookies", cookie_file_path])
-                            else:
-                                print(f"[palladium] cookie file missing, ignoring: {cookie_file_path}")
-
-                        sys.argv = [
-                            "yt-dlp",
-                            "-v",
-                            "--no-check-certificate",
-                            "--remote-components",
-                            "ejs:github",
-                            "--cache-dir",
-                            cache_dir if cache_dir else os.path.join(downloads_dir if downloads_dir else ".", ".cache"),
-                            "--force-overwrites",
-                            "--no-continue",
-                            "-P",
-                            run_output_dir if run_output_dir else ".",
-                            *output_args,
-                            *download_behavior_args,
-                            *preset_args,
-                            *extra_args,
-                            download_url,
-                        ]
-
-                        try:
-                            with patch_ytdlp_for_swiftffmpeg(bridge_adapter):
-                                runpy.run_module("yt_dlp", run_name="__main__", alter_sys=True)
-                            yt_exit_code = 0
-                        except KeyboardInterrupt:
-                            cancelled = True
-                            yt_exit_code = 130
-                            print("[palladium] yt-dlp cancelled by user")
-                        except SystemExit as exc:
-                            if exc.code is None:
-                                yt_exit_code = 0
-                            elif isinstance(exc.code, int):
-                                yt_exit_code = exc.code
-                            else:
-                                print(f"[palladium] unexpected SystemExit code: {exc.code}")
-                                yt_exit_code = 1
-                        except Exception:
-                            print("[palladium] yt-dlp execution failed")
-                            traceback.print_exc()
-                            yt_exit_code = 1
-
-                        if (
-                            not cancelled
-                            and auto_retry_failed_downloads
-                            and embed_thumbnail
-                            and download_playlist
-                            and playlist_progress.failed_item_records
-                        ):
-                            retry_candidates = playlist_progress.retry_candidates()
-                            for candidate in retry_candidates:
-                                item_index = candidate.get("index")
-                                item_title = candidate.get("title") or f"item {item_index}"
-                                print(f"[palladium] retrying playlist item without thumbnails: {item_title}")
-                                playlist_progress.set_tracking_suspended(True)
-                                retry_success, retry_error = run_retry_without_thumbnails(
-                                    retry_candidate=candidate,
-                                    download_url=download_url,
-                                    run_output_dir=run_output_dir,
-                                    cache_dir=cache_dir,
-                                    download_playlist=download_playlist,
-                                    output_args=output_args,
-                                    download_behavior_args=download_behavior_args,
-                                    preset_args=preset_args,
-                                    extra_args=extra_args,
-                                    bridge_adapter=bridge_adapter,
-                                )
-                                playlist_progress.set_tracking_suspended(False)
-                                if retry_success:
-                                    print(f"[palladium] retry without thumbnails succeeded: {item_title}")
-                                    playlist_progress.mark_retry_success(item_index)
-                                else:
-                                    print(f"[palladium] retry without thumbnails failed: {item_title}")
-                                    if retry_error:
-                                        print(retry_error)
-                                    playlist_progress.mark_retry_failed(item_index, retry_error)
-
-            if not cancelled and yt_exit_code is not None:
-                try:
-                    scan_dir = run_output_dir if run_output_dir else downloads_dir if downloads_dir else os.getcwd()
-                    downloaded_paths, primary_downloaded_path = detect_downloaded_files(scan_dir)
-                    if downloaded_paths:
-                        print(f"[palladium] downloaded files detected: {len(downloaded_paths)}")
-                        if primary_downloaded_path:
-                            print(f"[palladium] primary downloaded file: {primary_downloaded_path}")
-                        if yt_exit_code != 0:
-                            if has_primary_media_file(downloaded_paths):
-                                print(f"[palladium] overriding yt-dlp exit code {yt_exit_code} because a media file exists")
-                                yt_exit_code = 0
-                            else:
-                                print(f"[palladium] keeping yt-dlp exit code {yt_exit_code} because only sidecar files were downloaded")
-                    else:
-                        print("[palladium] downloaded files not detected")
-                except Exception:
-                    print("[palladium] unable to detect downloaded files")
-                    traceback.print_exc()
-        except KeyboardInterrupt:
-            cancelled = True
-            if yt_exit_code is None:
-                yt_exit_code = 130
-            print("[palladium] download flow cancelled by user")
-        except BaseException:
-            print("[palladium] unable to execute yt_dlp as __main__")
-            traceback.print_exc()
-            yt_exit_code = 1
-        finally:
-            sys.argv = argv_backup
-            try:
-                os.chdir(cwd_backup)
-            except Exception:
-                pass
-            if live_log_stream is not None:
-                try:
-                    live_log_stream.flush()
-                except Exception:
-                    pass
-
-        playlist_progress.finalize(cancelled=cancelled)
-        success = (pip_exit_code in (None, 0)) and (yt_exit_code == 0) and not cancelled
-        result_kind = playlist_progress.result_kind(success, cancelled)
-        if result_kind == "partial":
-            success = True
-        print(f"[palladium] flow success: {success}")
-
-    final_result_kind = playlist_progress.result_kind(success, cancelled)
-    return json.dumps({
-        "pip_attempted": pip_attempted,
-        "pip_exit_code": pip_exit_code,
-        "yt_exit_code": yt_exit_code,
-        "cancelled": cancelled,
-        "success": success,
-        "downloaded_paths": downloaded_paths,
-        "primary_downloaded_path": primary_downloaded_path,
-        "downloaded_path": primary_downloaded_path,
-        "output": output.getvalue(),
-        **playlist_progress.snapshot(result_kind=final_result_kind),
-    })
-
-
-def run_package_maintenance(action, custom_versions_json=None, live_log_fd_override=None, package_source_json_override=None):
-    output = TailBuffer()
-    console_stdout = sys.__stdout__ if sys.__stdout__ is not None else None
-    console_stderr = sys.__stderr__ if sys.__stderr__ is not None else None
-    pip_attempted = False
-    pip_exit_code = None
-    success = False
-    updates_available = False
-    updates_summary = "Not checked yet."
-    available_versions = {}
-    versions = {}
-    cancelled = False
-    restart_required = False
-    install_target = os.environ.get("PALLADIUM_PYTHON_PACKAGES")
-    cancel_file_path = os.environ.get("PALLADIUM_CANCEL_FILE", "").strip()
-    live_log_stream = open_live_log_stream(live_log_fd_override)
-    package_source = parse_package_source(package_source_json_override)
-
-    with contextlib.redirect_stdout(Tee(output, console_stdout, live_log_stream)), contextlib.redirect_stderr(Tee(output, console_stderr, live_log_stream)):
-        os.environ["PYTHONIOENCODING"] = "utf-8"
-        if install_target:
-            os.makedirs(install_target, exist_ok=True)
-            if install_target not in sys.path:
-                sys.path.insert(0, install_target)
-            print(f"[palladium] package install target: {install_target}")
-
-        try:
-            raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before start")
-            print(f"[palladium] package action: {action}")
-            print(f"[palladium] package source: {package_source.get('mode')}")
-            if action == "versions":
-                updates_available = False
-                updates_summary = "Skipped update check."
-                print("[palladium] quick version refresh only")
-            elif action == "index_versions":
-                updates_available = False
-                updates_summary = "Skipped update check."
-                available_versions = fetch_package_index_versions(
-                    install_target,
-                    allow_prereleases=bool(package_source.get("allow_prereleases")),
-                )
-                print("[palladium] fetched package index versions")
-            else:
-                updates_available, updates_summary = check_package_updates(install_target, package_source=package_source)
-                print(f"[palladium] updates available: {updates_available}")
-                print(f"[palladium] updates summary: {updates_summary}")
-
-            custom_versions = {}
-            if custom_versions_json:
-                try:
-                    parsed_versions = json.loads(custom_versions_json)
-                    if isinstance(parsed_versions, dict):
-                        for package_name in TRACKED_PACKAGES:
-                            raw_value = parsed_versions.get(package_name)
-                            if raw_value is None:
-                                continue
-                            requested_version = str(raw_value).strip()
-                            if requested_version:
-                                custom_versions[package_name] = requested_version
-                except Exception:
-                    print("[palladium] failed to parse custom version payload")
-                    traceback.print_exc()
-            if custom_versions:
-                print(f"[palladium] custom package versions requested: {custom_versions}")
-
-            if action in ("update", "reinstall"):
-                should_install = (
-                    action == "reinstall"
-                    or updates_available
-                    or bool(custom_versions)
-                    or package_source.get("mode") == "custom"
-                )
-                if should_install:
-                    raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before pip startup")
-                    pip_main = ensure_pip_entrypoint(install_target)
-                    if pip_main is not None:
-                        try:
-                            installed_versions = collect_versions(install_target=install_target, allow_cache_fallback=False)
-                            indexed_versions = fetch_package_index_versions(
-                                install_target=install_target,
-                                pip_main=pip_main,
-                                allow_prereleases=bool(package_source.get("allow_prereleases")),
-                            )
-                            if action == "reinstall":
-                                if package_source.get("mode") == "custom":
-                                    packages = list(package_source.get("custom_specs") or [])
-                                else:
-                                    packages = ["yt-dlp", "yt-dlp-apple-webkit-jsi"]
-                                cleanup_packages = ["yt-dlp", "yt-dlp-apple-webkit-jsi"]
-                            else:
-                                packages, cleanup_packages = build_package_install_plan(
-                                    installed_versions,
-                                    indexed_versions,
-                                    custom_versions=custom_versions,
-                                    package_source=package_source,
-                                )
-
-                            if not packages:
-                                print("[palladium] no package installs required")
-                                pip_exit_code = 0
-                            else:
-                                pip_attempted = True
-                                restart_required = invalidate_runtime_package_modules() or restart_required
-                                if install_target:
-                                    stale_removed = 0
-                                    for package_name in cleanup_packages:
-                                        raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled during cleanup")
-                                        stale_removed += cleanup_target_package(install_target, package_name)
-                                    print(f"[palladium] removed stale target package entries: {stale_removed}")
-                                raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before pip install")
-                                pip_args = build_pip_install_args(
-                                    packages,
-                                    install_target=install_target,
-                                    allow_prereleases=bool(package_source.get("allow_prereleases")),
-                                    upgrade=True,
-                                )
-                                pip_result = pip_main(pip_args)
-                                pip_exit_code = 0 if pip_result is None else int(pip_result)
-                                print(f"[palladium] pip exit code: {pip_exit_code}")
-                        except Exception:
-                            pip_exit_code = 1
-                            print("[palladium] pip update failed")
-                            traceback.print_exc()
-                    else:
-                        pip_exit_code = 1
-                else:
-                    print("[palladium] no updates available; skipping update")
-
-                raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before post-update check")
-                updates_available, updates_summary = check_package_updates(install_target, package_source=package_source)
-                print(f"[palladium] post-update updates available: {updates_available}")
-                print(f"[palladium] post-update updates summary: {updates_summary}")
-
-            if package_source.get("skip_webkit_patch"):
-                print("[palladium] skipping webkit patch by configuration")
-            else:
-                raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before webkit patch")
-                ensure_safe_webkit_jsi_runtime(install_target)
-
-            raise_if_cancel_requested(cancel_file_path, "[palladium] package action cancelled before version collection")
-            versions = collect_versions(install_target=install_target)
-            print(f"[palladium] yt-dlp version: {versions.get('yt-dlp')}")
-            print(f"[palladium] yt-dlp-apple-webkit-jsi version: {versions.get('yt-dlp-apple-webkit-jsi')}")
-            print(f"[palladium] gallery-dl version: {versions.get('gallery-dl')}")
-            print(f"[palladium] pip version: {versions.get('pip')}")
-
-            success = (pip_exit_code in (None, 0))
-            print(f"[palladium] package flow success: {success}")
-        except KeyboardInterrupt:
-            cancelled = True
-            success = False
-            print("[palladium] package action cancelled by user")
-
-    return json.dumps({
-        "pip_attempted": pip_attempted,
-        "pip_exit_code": pip_exit_code,
-        "success": success,
-        "cancelled": cancelled,
-        "updates_available": updates_available,
-        "updates_summary": updates_summary,
-        "versions": versions,
-        "available_versions": available_versions,
-        "restart_required": restart_required,
-        "output": output.getvalue(),
-    })
+from .gallery import run_gallery_dl_flow, run_gallery_dl_resolver
+from .maintenance import run_package_maintenance
+from .runtime import invalidate_runtime_package_modules
+from .ytdlp import run_yt_dlp_flow
+
+__all__ = [
+    "invalidate_runtime_package_modules",
+    "run_gallery_dl_flow",
+    "run_gallery_dl_resolver",
+    "run_package_maintenance",
+    "run_yt_dlp_flow",
+]
